backend/requirement.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
transformers==4.30.0
torch==2.0.1
requests==2.31.0